+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 
 ## Next release
 
+_Please stay tuned._
+
+
+## 1.0.2
+
 * Keeping attribute prefix while copying XML streams. [#23](https://github.com/difi/vefa-peppol/issues/23)
 * Updating model to allow for multiple process identifiers as part of one process metadata.
 * Adding concept of service reference in lookup.